+++ packages/capabilities/src/space/blob.js
@@ -11,9 +11,16 @@
  *
  * @module
  */
-import { equals as SpaceBlobCapabilities } from 'uint8arrays/equals'
+import { equals } from 'multiformats/bytes'
 import { capability, Schema, fail, ok } from '@ucanto/validator'
-import { equalBlob, equalWith, SpaceDID } from '../utils.js'
+import {
+  equalBlob,
+  equalWith,
+  SpaceDID,
+  and,
+  equal,
+  checkLink,
+} from '../utils.js'
 
 /**
  * Agent capabilities for Blob protocol
@@ -94,7 +101,7 @@ export const remove = capability({
       )
     } else if (
       delegated.nb.digest &&
-      !SpaceBlobCapabilities(delegated.nb.digest, claimed.nb.digest)
+      !equals(delegated.nb.digest, claimed.nb.digest)
     ) {
       return fail(
         `Link ${
@@ -160,7 +167,7 @@ export const get = capability({
       )
     } else if (
       delegated.nb.digest &&
-      !SpaceBlobCapabilities(delegated.nb.digest, claimed.nb.digest)
+      !equals(delegated.nb.digest, claimed.nb.digest)
     ) {
       return fail(
         `Link ${
@@ -170,6 +177,46 @@ export const get = capability({
     }
     return ok({})
   },
+})
+
+/**
+ * The `space/blob/replicate` capability allows an agent to replicate a Blob
+ * into a space identified by did:key in the `with` field.
+ *
+ * A replicate capability may only be invoked after a `blob/accept` receipt has
+ * been receieved, indicating the source node has successfully received the
+ * blob.
+ *
+ * Each Replicate task MUST target a different node, and they MUST NOT target
+ * the original upload target.
+ *
+ * The Replicate task receipt includes async tasks for `blob/replica/allocate`
+ * and `blob/replica/transfer`. Successful completion of the
+ * `blob/replica/transfer` task indicates the replication target has transferred
+ * and stored the blob. The number of `blob/replica/allocate` and
+ * `blob/replica/transfer` tasks corresponds directly to number of replicas
+ * requested.
+ */
+export const replicate = capability({
+  can: 'space/blob/replicate',
+  with: SpaceDID,
+  nb: Schema.struct({
+    /** Blob to replicate. */
+    blob: content,
+    /**
+     * The number of replicas to ensure. e.g. `replicas: 2` will ensure 3 copies
+     * of the data exist in the network.
+     */
+    replicas: Schema.integer().greaterThan(0),
+    /** Link to a location commitment indicating where the Blob must be fetched from. */
+    site: Schema.link({ version: 1 }),
+  }),
+  derives: (claimed, delegated) =>
+    and(equalWith(claimed, delegated)) ||
+    and(equalBlob(claimed, delegated)) ||
+    and(equal(claimed.nb.replicas, delegated.nb.replicas, 'replicas')) ||
+    and(checkLink(claimed.nb.site, delegated.nb.site, 'site')) ||
+    ok({}),
 })
 
 // ⚠️ We export imports here so they are not omitted in generated typedefs