+++ packages/capabilities/src/types.ts
@@ -18,10 +18,12 @@ import {
   ProofData,
   uint64,
 } from '@web3-storage/data-segment'
+import * as AssertCaps from './assert.js'
 import * as SpaceCaps from './space.js'
 import * as provider from './provider.js'
 import { top } from './top.js'
-import * as BlobCaps from './blob.js'
+import * as BlobCaps from './blob/index.js'
+import * as BlobReplicaCaps from './blob/replica/index.js'
 import * as SpaceBlobCaps from './space/blob.js'
 import * as W3sBlobCaps from './web3.storage/blob.js'
 import * as HTTPCaps from './http.js'
@@ -124,6 +126,20 @@ export interface DelegationNotFound extends Ucanto.Failure {
 }
 
 export type AccessConfirm = InferInvokedCapability<typeof AccessCaps.confirm>
+
+// Assert
+
+export type Assert = InferInvokedCapability<typeof AssertCaps.assert>
+export type AssertEquals = InferInvokedCapability<typeof AssertCaps.equals>
+export type AssertInclusion = InferInvokedCapability<
+  typeof AssertCaps.inclusion
+>
+export type AssertIndex = InferInvokedCapability<typeof AssertCaps.index>
+export type AssertLocation = InferInvokedCapability<typeof AssertCaps.location>
+export type AssertPartition = InferInvokedCapability<
+  typeof AssertCaps.partition
+>
+export type AssertRelation = InferInvokedCapability<typeof AssertCaps.relation>
 
 // Usage
 
@@ -530,6 +546,13 @@ export interface SliceNotFound extends Failure {
 export type Blob = InferInvokedCapability<typeof BlobCaps.blob>
 export type BlobAllocate = InferInvokedCapability<typeof BlobCaps.allocate>
 export type BlobAccept = InferInvokedCapability<typeof BlobCaps.accept>
+export type BlobReplica = InferInvokedCapability<typeof BlobReplicaCaps.replica>
+export type BlobReplicaAllocate = InferInvokedCapability<
+  typeof BlobReplicaCaps.allocate
+>
+export type BlobReplicaTransfer = InferInvokedCapability<
+  typeof BlobReplicaCaps.transfer
+>
 export type SpaceBlob = InferInvokedCapability<typeof SpaceBlobCaps.blob>
 export type SpaceBlobAdd = InferInvokedCapability<typeof SpaceBlobCaps.add>
 export type SpaceBlobRemove = InferInvokedCapability<
@@ -537,6 +560,9 @@ export type SpaceBlobRemove = InferInvokedCapability<
 >
 export type SpaceBlobList = InferInvokedCapability<typeof SpaceBlobCaps.list>
 export type SpaceBlobGet = InferInvokedCapability<typeof SpaceBlobCaps.get>
+export type SpaceBlobReplicate = InferInvokedCapability<
+  typeof SpaceBlobCaps.replicate
+>
 /** @deprecated */
 export type W3sBlob = InferInvokedCapability<typeof W3sBlobCaps.blob>
 /** @deprecated */
@@ -597,6 +623,42 @@ export interface SpaceBlobGetSuccess extends BlobItem {}
 // TODO: make types more specific
 export type SpaceBlobGetFailure = Ucanto.Failure
 
+// Blob replicate
+export interface SpaceBlobReplicateSuccess {
+  site: UCANAwait<'.out.ok.site'>[]
+}
+
+/** Too many or too few replicas were instructed. */
+export interface ReplicationCountRangeError extends Failure {
+  name: 'ReplicationCountRangeError'
+}
+
+/** There are not enough replication nodes available to replicate the data. */
+export interface ReplicationCandidateUnavailable extends Failure {
+  name: 'ReplicationCandidateUnavailable'
+}
+
+/** Blob to replicate was not found in the space. */
+export interface ReplicationSourceNotFound extends Failure {
+  name: 'ReplicationSourceNotFound'
+}
+
+/**
+ * The location commitment was invalid in some way. For example, it has expired,
+ * is revoked, had a signature that did not verify or referenced a blob that was
+ * not requested to be replicated.
+ */
+export interface InvalidReplicationSite extends Failure {
+  name: 'InvalidReplicationSite'
+}
+
+export type SpaceBlobReplicateFailure =
+  | ReplicationCountRangeError
+  | ReplicationCandidateUnavailable
+  | ReplicationSourceNotFound
+  | InvalidReplicationSite
+  | Failure
+
 // Blob allocate
 export interface BlobAllocateSuccess {
   size: number
@@ -630,6 +692,22 @@ export interface AllocatedMemoryHadNotBeenWrittenTo extends Ucanto.Failure {
 export type BlobAcceptFailure =
   | AllocatedMemoryHadNotBeenWrittenTo
   | Ucanto.Failure
+
+// Blob replica allocate
+export interface BlobReplicaAllocateSuccess {
+  size: number
+  site: UCANAwait<'.out.ok.site'>
+}
+
+export type BlobReplicaAllocateFailure = Failure
+
+// Blob replica transfer
+export interface BlobReplicaTransferSuccess {
+  // A Link for a delegation with location commitment for the replicated blob.
+  site: Link
+}
+
+export type BlobReplicaTransferFailure = Failure
 
 // Storage errors
 export type StoragePutError = StorageOperationError
@@ -922,6 +1000,13 @@ export type ServiceAbility = TupleToUnion<ServiceAbilityArray>
 
 export type ServiceAbilityArray = [
   Top['can'],
+  Assert['can'],
+  AssertEquals['can'],
+  AssertInclusion['can'],
+  AssertIndex['can'],
+  AssertLocation['can'],
+  AssertPartition['can'],
+  AssertRelation['can'],
   ProviderAdd['can'],
   Space['can'],
   SpaceInfo['can'],
@@ -970,11 +1055,15 @@ export type ServiceAbilityArray = [
   Blob['can'],
   BlobAllocate['can'],
   BlobAccept['can'],
+  BlobReplica['can'],
+  BlobReplicaAllocate['can'],
+  BlobReplicaTransfer['can'],
   SpaceBlob['can'],
   SpaceBlobAdd['can'],
   SpaceBlobRemove['can'],
   SpaceBlobList['can'],
   SpaceBlobGet['can'],
+  SpaceBlobReplicate['can'],
   W3sBlob['can'],
   W3sBlobAllocate['can'],
   W3sBlobAccept['can'],