+++ packages/upload-api/src/test/lib.js
@@ -5,7 +5,10 @@ import * as AdminUploadInspect from './handlers/admin/upload/inspect.js'
 import * as RateLimitAdd from './handlers/rate-limit/add.js'
 import * as RateLimitList from './handlers/rate-limit/list.js'
 import * as RateLimitRemove from './handlers/rate-limit/remove.js'
-import * as Blob from './handlers/blob.js'
+import * as BlobAdd from './handlers/space/blob/add.js'
+import * as BlobList from './handlers/space/blob/list.js'
+import * as BlobRemove from './handlers/space/blob/remove.js'
+import * as BlobReplicate from './handlers/space/blob/replicate.js'
 import * as Ucan from './handlers/ucan.js'
 import * as Subscription from './handlers/subscription.js'
 import * as Upload from './handlers/upload.js'
@@ -17,12 +20,22 @@ import { test as agentStoreTests } from './storage/agent-store-tests.js'
 import { test as delegationsStorageTests } from './storage/delegations-storage-tests.js'
 import { test as provisionsStorageTests } from './storage/provisions-storage-tests.js'
 import { test as rateLimitsStorageTests } from './storage/rate-limits-storage-tests.js'
+import { test as replicaStorageTests } from './storage/replica-storage-tests.js'
 import { test as revocationsStorageTests } from './storage/revocations-storage-tests.js'
 import { test as plansStorageTests } from './storage/plans-storage-tests.js'
 import { DebugEmail } from '../utils/email.js'
 export * as Context from './helpers/context.js'
 
 export * from './util.js'
+
+const Blob = {
+  test: {
+    ...BlobAdd.test,
+    ...BlobList.test,
+    ...BlobRemove.test,
+    ...BlobReplicate.test,
+  },
+}
 
 export const test = {
   ...Blob.test,
@@ -35,6 +48,7 @@ export const storageTests = {
   ...delegationsStorageTests,
   ...provisionsStorageTests,
   ...rateLimitsStorageTests,
+  ...replicaStorageTests,
   ...revocationsStorageTests,
   ...plansStorageTests,
   ...blobRegistryTests,
@@ -66,6 +80,7 @@ export {
   delegationsStorageTests,
   provisionsStorageTests,
   rateLimitsStorageTests,
+  replicaStorageTests,
   revocationsStorageTests,
   plansStorageTests,
   blobRegistryTests,