+++ packages/upload-api/test/handlers/space/blob/remove.spec.js
@@ -1,0 +1,4 @@
+import { test } from '../../../../src/test/test.js'
+import * as BlobRemove from '../../../../src/test/handlers/space/blob/remove.js'
+
+test({ 'space/blob/remove': BlobRemove.test })