+++ packages/upload-api/test/storage/replica-storage.spec.js
@@ -1,0 +1,3 @@
+import * as ReplicaStorage from '../../src/test/storage/replica-storage-tests.js'
+import { test } from '../../src/test/test.js'
+test({ 'in memory replica storage': ReplicaStorage.test })