+++ packages/encrypt-upload-client/CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.0.26 (2025-05-22)
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/upload-client to 1.2.2
+- Updated @storacha/capabilities to 1.6.0
+- Updated @storacha/client to 1.3.2
+
+## 0.0.19 (2025-05-21)
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/upload-client to 1.2.1
+- Updated @storacha/capabilities to 1.5.0
+- Updated @storacha/client to 1.3.0
+
+## 0.0.12 (2025-05-13)
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/upload-client to 1.2.0
+- Updated @storacha/client to 1.2.9
+
 ## 0.0.11 (2025-05-07)
 
 ### 🧱 Updated Dependencies