+++ packages/ui/packages/core/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 2.4.59 (2025-05-22)
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/access to 1.2.1
+- Updated @storacha/client to 1.3.2
+
+## 2.4.52 (2025-05-21)
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/access to 1.2.0
+- Updated @storacha/client to 1.3.0
+
+## 2.4.45 (2025-05-13)
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/client to 1.2.9
+
 ## 2.4.43 (2025-05-07)
 
 ### 🧱 Updated Dependencies