+++ packages/upload-client/CHANGELOG.md
@@ -1,3 +1,29 @@
+## 1.2.2 (2025-05-22)
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/filecoin-client to 1.0.8
+- Updated @storacha/capabilities to 1.6.0
+- Updated @storacha/blob-index to 1.0.7
+
+## 1.2.1 (2025-05-21)
+
+### 🧱 Updated Dependencies
+
+- Updated @storacha/filecoin-client to 1.0.7
+- Updated @storacha/capabilities to 1.5.0
+- Updated @storacha/blob-index to 1.0.6
+
+## 1.2.0 (2025-05-13)
+
+### 🚀 Features
+
+- perf: faster small file uploads ([#257](https://github.com/storacha/upload-service/pull/257))
+
+### ❤️ Thank You
+
+- ash
+
 ## 1.0.9 (2025-03-27)
 
 ### 🧱 Updated Dependencies